type: FileInput 上传输入框修复类型提示
