+++ packages/react-file-input/src/Input.tsx
@@ -1,11 +1,8 @@
 import React from 'react';
-import { InputProps } from '@uiw/react-input';
-import { FileInputProps } from './';
+import { FileInputDefaultProps } from './types';
 import { FileInputStyleWarp } from './style';
 
-export interface InputUploadProps extends FileInputProps, InputProps {}
-
-export default React.forwardRef<HTMLInputElement, InputUploadProps>((props, ref) => {
+export default React.forwardRef<HTMLInputElement, FileInputDefaultProps>((props, ref) => {
   const { className, dataLabel = 'Browse', prefixCls = 'w-fileinput', ...other } = props;
   const cls = [prefixCls, className].filter(Boolean).join(' ').trim();
   return <FileInputStyleWarp ref={ref} data-label={dataLabel} className={cls} {...other} type="file" />;
