+++ packages/react-file-input/src/types.ts
@@ -1,7 +1,21 @@
-import { FileInputValue } from './index';
+import { InputProps } from '@uiw/react-input';
 
-export type UploadType = 'input' | 'picture' | 'text' | 'card';
+export type UploadType = 'input' | 'picture' | 'text' | 'card' | undefined;
 
+// type IItemGeneric<T, U> = Exclude<{
+//   uploadType?: T;
+// }, U>
+
+type IItemGeneric<T, U> = {
+  uploadType?: T;
+} & Omit<U, 'uploadType'>;
+
+export interface FileInputValue {
+  dataURL?: string;
+  file?: File;
+  name?: string;
+  [key: string]: any;
+}
 export interface FileInputShowIconProps {
   showPreviewIcon?: boolean;
   showRemoveIcon?: boolean;
@@ -9,10 +23,9 @@ export interface FileInputShowIconProps {
 }
 
 export interface FileInputStyleBaseProps {
+  uploadType?: UploadType;
   className?: string;
   prefixCls?: string;
-  /** 上传列表的内置样式 */
-  uploadType?: UploadType;
   multiple?: boolean;
 }
 
@@ -28,3 +41,19 @@ export interface FileInputUploadProps extends FileInputStyleBaseProps {
   /** 预览 */
   onPreview?: (file: FileInputValue) => void;
 }
+
+export interface FileInputDefaultProps extends FileInputStyleBaseProps, InputProps {
+  dataLabel?: string;
+}
+export interface FileInputListProps extends FileInputUploadProps {
+  shape?: 'circle' | 'round';
+  size?: 'large' | 'middle' | 'small' | 'default';
+  showFileIcon?: FileInputShowIconProps;
+  dataList?: FileInputValue[];
+  onAdd?: () => void;
+  onRemove?: (index: number) => void;
+}
+
+export type FileInputProps =
+  | IItemGeneric<'input' | undefined, FileInputDefaultProps>
+  | IItemGeneric<'picture' | 'text' | 'card', FileInputListProps>;
