+++ website/src/components/Nav/index.tsx
@@ -140,7 +140,7 @@ export default function Nav() {
         />
       </div>
       <div className={[styles.btn, state.layout === 'left' ? null : styles.btnTop].filter(Boolean).join(' ').trim()}>
-        <dark-mode light="Light" dark="Dart" ></dark-mode>
+        {/* <dark-mode light="Light" dark="Dart" ></dark-mode> */}
         <select value={i18n.language} onChange={(e) => changeLanguage(e)}>
           <option value="zh-CN">简体中文</option>
           <option value="en-US">English</option>
